Refactoring clock system
Created a clock struct with all the ticks and latches etc... required for stuff to run.
The Virtual_frame_buffer uses one to blink character for ex.
The main uses one for timing stuff, independent from the frame buffer.
Went back to event Polling with V-Sync so everything runs at the framerate of the display ...